+++ apps/frontend/src/components/CropMarket/AskingPrice.tsx
