+++ apps/frontend/src/components/CropMarket/Chart.tsx
