+++ apps/frontend/src/components/CropMarket/Trade.tsx
