+++ apps/frontend/src/components/CropMarket/TradeSection.tsx
