+++ apps/frontend/src/components/Icons/CloseIcon.tsx
