+++ apps/frontend/src/components/Intro/IntroTitle.tsx
