+++ apps/frontend/src/components/Lottery/LotteryIcon.tsx
