+++ apps/frontend/src/components/Lottery/LotteryModal.tsx
