+++ apps/frontend/src/components/Lottery/LotteryTicket.tsx
