+++ apps/frontend/src/components/public/Footer.tsx
